render posts on profile page
