+++ src/components/Posts/postView.tsx
@@ -1,0 +1,38 @@
+import dayjs from "dayjs";
+import relativeTime from "dayjs/plugin/relativeTime";
+import Image from "next/image";
+import Link from "next/link";
+import type { RouterOutputs } from "~/utils/api";
+
+type PostWithUser = RouterOutputs["post"]["getAll"][number];
+
+dayjs.extend(relativeTime);
+
+const PostView = ({ post, author }: PostWithUser) => {
+  return (
+    <div key={post.id} className="flex  gap-3 border-b border-slate-400 p-8">
+      <Image
+        src={author.profileImageUrl}
+        className="h-6 w-6 rounded-full"
+        alt={`@${author.username}'s profile image`}
+        width="56"
+        height="56"
+      />
+      <div className="flex flex-col">
+        <div className="flex gap-1 text-slate-300">
+          <Link href={`/@${author.username}`}>
+            <span>{`@${author.username} `}</span>
+          </Link>
+          <Link href={`/post/${post.id}`}>
+            <span className="font-thin">{` . ${dayjs(
+              post.createdAt
+            ).fromNow()}`}</span>
+          </Link>
+        </div>
+        <span className="text-xl">{post.content}</span>
+      </div>
+    </div>
+  );
+};
+
+export default PostView;
