+++ src/pages/[slug].tsx
@@ -4,6 +4,22 @@ import Image from "next/image";
 import PageLayout from "~/layouts/page";
 import { api } from "~/utils/api";
 
+const ProfileFeed = ({ userId }: { userId: string }) => {
+  const { data, isLoading } = api.post.getPostByUserId.useQuery({ userId });
+
+  if (isLoading) return <LoadingPage />;
+
+  if (!data || data.length === 0) return <div>User has not posted</div>;
+
+  return (
+    <div className="flex flex-col">
+      {data.map((fullPost) => {
+        return <PostView key={fullPost.post.id} {...fullPost} />;
+      })}
+    </div>
+  );
+};
+
 const ProfilePage: NextPage<{ username: string }> = ({ username }) => {
   const { data } = api.profile.getUserByUsername.useQuery({
     username,
@@ -30,6 +46,7 @@ const ProfilePage: NextPage<{ username: string }> = ({ username }) => {
           data.username ?? ""
         }`}</div>
         <div className="w-full border-b border-slate-400"></div>
+        <ProfileFeed userId={data.id} />
       </PageLayout>
     </>
   );
@@ -39,6 +56,8 @@ import { createProxySSGHelpers } from "@trpc/react-query/ssg";
 import superjson from "superjson";
 import { appRouter } from "~/server/api/root";
 import { prisma } from "~/server/db";
+import { LoadingPage } from "~/components/Loading";
+import PostView from "~/components/Posts/postView";
 
 export const getStaticProps: GetStaticProps = async (context) => {
   const ssg = createProxySSGHelpers({
