+++ src/pages/index.tsx
@@ -1,18 +1,12 @@
 import { SignInButton, useUser } from "@clerk/nextjs";
 import { type NextPage } from "next";
 import Image from "next/image";
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/relativeTime";
-
-import { api } from "~/utils/api";
-import type { RouterOutputs } from "~/utils/api";
-import { LoadingComponent, LoadingPage } from "~/components/Loading";
 import { useState } from "react";
 import toast from "react-hot-toast";
-import Link from "next/link";
+import { LoadingComponent, LoadingPage } from "~/components/Loading";
+import PostView from "~/components/Posts/postView";
 import PageLayout from "~/layouts/page";
-
-dayjs.extend(relativeTime);
+import { api } from "~/utils/api";
 
 const CreatePostWizard = () => {
   const { user } = useUser();
@@ -67,35 +61,6 @@ const CreatePostWizard = () => {
   );
 };
 
-type PostWithUser = RouterOutputs["post"]["getAll"][number];
-
-const PostView = (props: PostWithUser) => {
-  const { post, author } = props;
-  return (
-    <div key={post.id} className="flex  gap-3 border-b border-slate-400 p-8">
-      <Image
-        src={author.profileImageUrl}
-        className="h-6 w-6 rounded-full"
-        alt={`@${author.username}'s profile image`}
-        width="56"
-        height="56"
-      />
-      <div className="flex flex-col">
-        <div className="flex gap-1 text-slate-300">
-          <Link href={`/@${author.username}`}>
-            <span>{`@${author.username} `}</span>
-          </Link>
-          <Link href={`/post/${post.id}`}>
-            <span className="font-thin">{` . ${dayjs(
-              post.createdAt
-            ).fromNow()}`}</span>
-          </Link>
-        </div>
-        <span className="text-xl">{post.content}</span>
-      </div>
-    </div>
-  );
-};
 const Feed = () => {
   const { data, isLoading: postsLoading } = api.post.getAll.useQuery();
